+++ mindone/transformers/mindspore_adapter/paged_attention_infer_attention_block.py
@@ -6,7 +6,7 @@ from mindspore import ops
 from mindspore.common.tensor import Tensor
 from mindspore.nn.cell import Cell
 
-from .flash_attention import FlashAttention
+from .paged_attention_fa_block import FlashAttention
 from .paged_attention_mgr import PagedAttentionMgr
